Extract embed link code to separate functions
